ajuste para borrar libro y los comentarios asociados a él

+++ api/src/routes/controllers/books/bookDelete.js
@@ -1,12 +1,34 @@
+import { jwt } from 'jsonwebtoken'
+
 import Book from '../../../models/Book.js'
+import Comment from '../../../models/Comment.js'
 
 const bookDelete = async (req, res, next) => {
   try {
     const { id } = req.params
 
-    if (!id) res.status(200).json('ID required')
-    await Book.findByIdAndDelete(id)
-    res.status(200).json(`The workbook with the ID ${id} was deleted`)
+    if (!id) res.status(400).json('ID required')
+
+    const authorization = req.get('authorization')
+
+    if (authorization.length <= 7) {
+      res.status(401).json('token missing')
+    }
+    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+      const token = authorization.substring(7)
+      const decodedToken = jwt.verify(token, process.env.SECRET)
+
+      if (!token || !decodedToken.id) {
+        return res.status(401).json({ error: 'token missing or invalid' })
+      }
+      const book = await Book.findById(id)
+
+      await Comment.deleteMany({ _id: { $in: book.comment } })
+
+      const deletedBook = await Book.findByIdAndDelete(id)
+
+      res.status(200).json(`The workbook  ${deletedBook.title} was deleted`)
+    }
   } catch (error) {
     next(error)
   }
